Notifications: Settings updates and some code moving.
Change notification settings to save when the user leaves the view so
that we're not over-firing the xml-rpc save call. Change the error
dialog to show the returned xml-rpc error instead of a generic error
message.

Also renamed the sendPushNotificationsBlogsList method to
'syncPushNotificationInfo' since we are doing more than just sending
the blogs list. Also moved the method to WordPressComApi, since it
makes sense to have it live there vs. the app delegate.